Improved error messages in Registry
Inside a large application with many moving parts, the message "Not registered" is a bit confusing.

Including the extra "factory" in there is just enough context to understand that the developer
has borked something in FactoryGirl and not elsewhere

Also, the error in `add_as` was using different nomenclature; changed that to be consistent and also use "factory" in the message

I don't normally test for exact error messages, but since you had one such in the acceptance
specs, I added two tests to the factory_girl/registry spec that look for this exact error message.