Add an alphabetic_sequence method
Use `alphabetic_sequence(:name) { |n| "=#{n}" }` and all values will be
alphabetically sorted. This solves the issue when writing tests that
rely on alphabetical sorting and the factories cross a string length
boundary (e.g. "Z".."AA" ("Z" < "AA") or "9".."10") will often pass
in isolation, but when run in a full test suite might fail intermittently.

If your tests ever reach a point where the next value is `<` the previous
value, an error will be raised.