Raise an error when a sequence returns a value that is less than the previous
The prime example is:

`sequence(:name) { |n| "name#{n}" }`

If your code evers sort by `name` alphabetically, then you could end up
with the case where your sequence returns `name9`, `name10`. The test
might pass in isolation, but when run as part of a full suite, you could
end up with intermittent failures based on the ordering of the specs.

This raises an error if a value is ever less than (`<`) the previous value.
The solution is to change your sequence to `sequence(:name, "00000000")`
(or some longer string if you call the sequence several millions of times).

If you are using an enumerator, this check is skipped.