Test all currently-disabled persistence methods on Stub strategy
Many of the methods defined in ActiveRecord::Persistence are disabled in the
Stub strategy in order to prevent stub instances from accessing the database.
Tests exist for some, but not all of these overrides.

This change ensures that all the currently-overridden methods have matching
tests. In a future commit this list will be expanded to disable all methods
in ActiveRecord's persistence API that aren't given stub implementations by
the Stub strategy.