Refactor disabled persistence methods to blacklist constant
Previously, a number of persistence methods were disabled on instances created
by the Stub strategy, in order to prevent stubbed instances from talking to
the database.

This refactors these method definitions to use a blacklist of method names,
rather than individually overriding each method. As a result all overridden
methods now accept arbitrary arguments. I think this is justifiable, as we
are, after all, disabling these methods - it shouldn't matter on a stub
instance whether the caller is calling them correctly, since on the instances
this code is responsible for, and therefore in any code that refers to them,
the methods should not be called at all.