More intuitive attribute assignment
When building a factory, self in attribute blocks feels like it should
somehow be related to the instance. This commit allows self to be passed
like the instance. This allow for a more fluid definition of factories
in which you can pass self, and then on the other side access attributes
like they already exist.

In the spec you can see that the comment factory uses the post title,
despite the post title being in a delayed execution block which I think
is a huge deal for enabling simpler factories and could allow for 95% of
callback related code to be deleted.