clarify project name change links in README
The README referred to NAME.md in the old repository, which now redirects to the new repository, and had a separate "Interested in the name change?" link which pointed to the same NAME.md file in the current repository -- confusing! This commit copyedits that section to more clearly link to both the old explanation and the github issue thread which initially discussed changing the name.