Revert "Replace instructions with link to guide"
This reverts commit 2fb828b0df29ef1d2eaf974340fbb048074194ee.

A link to a document referred to in a tagged commit does not allow for changes and upgrades made to that particular document. Instructions are a living/breathing document that are subject to changes and improvements. I had originally wanted to improve on the upgrade instructions but could not do so readily, because they were linked to a particular tagged commit. Hence had to revert the change to allow for upgrades to be made. I have incorporated the entire changes in the linked document herein. If a link is to be made, then it must be to a document in the current head, not a fixed commit, so that changes can be readily made.