Simplify RSpec configuration instructions
As noted in the original PR #1105, separating the Rails and non-Rails
configurations makes it easier to copy and paste.

We are also removing the bit about requiring `support/factory_bot`. I
think think it is outside the scope of this documentation to suggest
splitting of RSpec configuration into multiple files. We don't specify
file names for the other test configurations. Also, this line has
led to more confusion than it is probably worth, and has come up
in several PRs: #987, #1025, #1081, #1085, and #1091.

Co-authored-by: Jarvis Johnson <jarviscjohnson@gmail.com>
Co-authored-by: Daniel Colson <danieljamescolson@gmail.com>