Inherit from BasicObject instead of undefining
`FactoryBot::DefinitionProxy` currently undefines all but a few methods.
Instead, it can inherit from `BasicObject` and then get any other methods
it needs (`#block_given?` and `#raise` are it as far as I can tell)
directly from `Kernel`.

This way if somebody adds a method to Kernel after loading a factory
that has an attribute of the same name as that method, there is still
some hope their code might work. (Of course nobody is ever adding
methods to Kernel, right?)