Update documentation for adding attributes
I am leaving the Static Attributes section in the README for now, since
it might still be helpful for people upgrading, but I took out the
example because I don't want to show invalid code.

I removed reference to static attributes in the definition_proxy
documentation, and updated the signature of add_attributes. I also
got rid of the bit about yielding a FactoryBot::Strategy, which is not
correct. We document correctly in the GETTING_STARTED that we yield an
evaluator instance.