Add rubocop
to run rubocop locally and not to wait for HoundCI to bark at us.

The version is locked to 0.54 because this is the version HoundCI is using.
Using a newer version will not work as config changed from 0.54 to currently 0.59
and rubocop would complain with the old configuration.

The configuration is using the default configuration from version 0.54.

See http://help.houndci.com/configuration/rubocop.

Fix #1195.