WIP - aliases interfere with fields ending in _id with no association
If a field name ends in _id and isn't intended for an association,
the default FactoryBot.aliases interfere.

There doesn't seem to be a clear way to fix this problem.  One idea
would be to have a way in a factory definition to mark an attribute
to not be aliased.

The committed spec will fail with the default aliases.  The user can
always override this with the following to get tests passing:

`FactoryBot.aliases = [[/^((?!item_type).)_id/, '\1']]`

This is feasible but not an ideal situation.