Remove unhelpful lint specs
These specs have no expectations, and they are passing
`validate_traits` as an option, which is not an option the linter
understands.

Since the first spec is already covered by "allows for selective
linting" above, and what I assume the second spec was meaning to cover
is covered by the trait specs, I see no reason to keep these around.