Test all combinations of indifferent trait access
We were only explicitly testing defining a trait with a string then
referencing it with either a symbol or a string. We were implicitly
testing defining a trait with a symbol then referencing it with a
symbol. We were not testing defining a trait with a symbol then
referencing it with a string.