Raise KeyError instead of ArgumentError in Registry
to make use of did_you_mean gem.

The did_you_mean gem only supports NameError, NoMethodErros
and KeyError. However, for NameError the message does also need
to match a certain format and we can not use a custom message
like 'Factory not registered ...'. Therefore using KeyError
is to only logical conclusion.

Additional the did_you_mean gem makes use of the receiver attributes.
In Ruby > 2.5 it is not possible to set the receiver and key attributes
on a KeyError. Therefore it was also necessary to monkey patch the
KeyError class. Starting in Ruby 2.6 this will not be necessary anymore
https://bugs.ruby-lang.org/issues/14313.

Fix #992

Co-authored-by: Daniel Colson <danieljamescolson@gmail.com>