Pass on the index of the created object to a block when using *_list methods
If a block is provided to a call to any of the `*_list` methods, pass
the created object's index along with the object in the arguments of the
block.

A block that only takes in the object is still allowed and works.

A block that takes in both the object and its index is allow and works.