feature Add file_fixture helper to syntax methods
Why:
This is to add simple support to files on factories. This is the
original [issue](https://github.com/thoughtbot/factory_bot/issues/1282).

What:
It is a common behavior to interact with files when creating factories,
and it is used heavily, as demonstrated with the many examples in the
issue, this is a first draft and I would love to see more use-cases to
add to the acceptance test. For the time being we have at least the base
case added.