Use already installed version of bundler
These before install steps were introduced quite a while back in
850116da, 3c0547fa, and b3509f81. It looks like they were introduced as
workarounds to no-longer-relevant issues. Now that bundler is included
in Ruby by default, this is causing builds to fail while waiting on a
prompt about where it is OK to overwrite the "bundle" executable.

We still need to install bundler for the jruby builds, so we can go ahead
and install it if no `bundle` executable is available.