Rough draft of Namespaced Factories
See: https://github.com/thoughtbot/factory_bot/issues/199

I have a number of clients who use FactoryBot and are working through
the difficult challene of decomposing a large monolith into either
domain-driven namespaces; or small gems/engines that are "plugged into"
a larger monolith.

In most cases, we've been able to get away with keeping the factories
bundled with their extracted modules and using distinct names.

However there's also been a bit of copy-pasting of factory code back into the monolith in cases where the names really _do_ need to be shared.

This is an attempt to bring @joshuaclayton's example of namespacing into
the core FactoryBot DSL.

If you'd prefer I packaged this as it's own gem, I'd be happy to do so;
but I'd be stoked to see this in the core FactoryBot package!