Replace Travis with GitHub Actions
Why:
----

Github actions have the added benefit of being integrated in the same platform we are using and it is one less place to look for things. As an added bonus, the performance and speed of the test suite has picked up a bit and while that is not a big requirement for our library it is an added benefit 🥳 

How:
----

The travis CI setup is replaced by a github workflow making use of the same folder to run the tests against a varied array of versions of ruby and rails. This pattern helps us to have a solid Gemfile to point to when making local tests for regression.

Co-authored-by: Alejandro Dustet <aledustet@gmail.com>