Document 6.0.0 may break FactoryBot.lint
In a sufficiently large codebase, the automatic definition of traits for
Active Record enum may result in more combinations being checked by
`FactoryBot.lint`. This breaking change may not be obvious because the
first impression of the error messages may be "these factories were
passing before the upgrade but now failing ?", while the truth is that
they don't exist before the upgrade.