Don't run simplecov on alternate Ruby
The truffleruby + Rails 7.0 build was failing because of some kind of
simplecov incompatibility. I don't think we need to run simplecov on the
truffleruby and jruby builds (we might not need it on any build—it
prints some output but doesn't affect the build status at all—but I
figured I'd leave it alone for now in case somebody is looking at it).

I needed to add a fileutils require to one test that was presumably
getting implicitly required before via simplecov.