Move documentation into mdBook
The `GETTING_STARTED.md` file has becoming increasingly hard to
navigate, so split it into a book format. This uses the popular mdBook
program to generate nice HTML from Markdown files.

It also adds a GitHub Action to deploy an updated book when merging a
branch into main.

Barely any editing has been done to the documentation itself; the goal
was to port it over and then improve it separately. (But I did remove
"just" and "please" in some places.)