Major docs overhaul: refs, guide, cookbook
Split the docs into reference, guide, and cookbook.

The guide is mostly as it was before, with a lot of cleanup. Some
sections were combined and simplified.

Some guides were moved into a new cookbook section. This section
describes techniques that are helpful in specific situations of using
factory_bot but not, in itself, part of factory_bot. For example, using
factory_bot to simulate a has_many relation.

Other guides were moved into the wiki. The criteria for this was: does
it have to do with factory_bot, or with something external? Will we
change the docs alongside a change in this repo's code, or due to
another repo's code change?

A new reference section is added. It lists terse facts about methods you
are expected to use. The reason RubyDoc isn't the right solution here is
because these methods are defined all over the codebase. Due to
`instance_eval` and mixins, factory_bot is more of a DSL than a set of
Ruby classes. Thus, the reference section.

During the review, we updated the name of the second parameter for
hooks: instead of `evaluator`, we call it the `context`.

Co-authored-by: Sara Jackson <csara@thoughtbot.com>
Co-authored-by: Stefanni Brasil <stefannibrasil@gmail.com>