Reject `sequence` definitions for Active Record primary keys
It might be useful to define `sequence(:id)` attributes for models
representing objects in external systems. It might be tempting to apply
that same pattern to Active Record-backed models that are internal to
the system under test.

However, a factory that generates its own sequence of `id` (or any other
primary key attribute) risks collision or desynchronization with the
underlying datastore (most likely an SQL database with its own primary
key generation strategy and "next" value).

This commit introduces a check in the `FactoryBot::Definition#compile`
method to ensure that descendants of `ActiveRecord::Base` cannot declare
a sequence to generate their own primary key.