Trigger update-readme workflow
The workflow had `[ci-skip]` in the commit. But some repos require all CI actions to be ran for a commit. So we need to run CI checks. I updated the templates workflow to not skip ci anymore. That means we need to trigger the workflow again with the changes and not just re-run the failed action run because it was still calling `[ci-skip]`.