Update test matrix dependencies
We have some outdated dependencies, and are seeing failures on the CI
environment for Rails 6.1.

The Rails 6.1 failures happens while resolving the dependencies, and I
suspect that is happening because the Rails 6.1 Gemfile got messed up by
Dependabot and is in an inconsistent state. I haven't gone deep into
that, but this is an easy (and necessary) experiment.