Limit `push` builds to the `main` branch
GitHub Actions has been running the lint twice in Pull Requests: once
following the `on push` trigger, once following `on pull_request`. This
duplication is currently avoided when running tests by an `if` rule,
which skips if the event isn't a push or the Pull Request doesn't come
from a fork.

This commit removes the `if` rule, and limits the `push` builds to the
`main` branch. That will ensure that merged Pull Requests will trigger a
build, and that any Pull Request activity will also trigger a build.
Being a global rule for the whole build, `standard` will not run twice
anymore.