fix: addresses regression in attribute assignment
factory_bot v6.5.5 included a patch which worked around the previous
limitaion of not being able to assign both <attribute> and
<attribute>_id as independent attributes. The change introduced
regressions into how attributes are assigned.

This commit fixes regressions involving declared attributes which are
aliases of overrides where an association is involved. It fixes behavior
where the attributes are not assigned in the expected manner as well as
fixes an instance where an extra erroneous associated object is created.